added users admin roles
Added a static currently logged in user called loggedInUser added roles to the user class and edited saving and loading to reflect them. Did a check in the main window to  when clicking on manage users to see if you are on an "administrator" user and allow access to that screen.